requirements.txt
pytesseract 
pdf2image 
ingua-language-detector 
python-docx 
pypdf